Refactor aggregate import from libs.data to libs.domain for consistency
