+++ libs/commands/results/summary.py
@@ -5,7 +5,7 @@ libs/commands/results/summary.py
 from typing import TYPE_CHECKING
 
 import libs.global_value as g
-from libs.data import aggregate
+from libs.domain import aggregate
 from libs.domain.datamodels import GameInfo
 from libs.functions import message
 from libs.types import StyleOptions
