+++ libs/data/__init__.py
@@ -1,7 +1,6 @@
 """
 データ操作モジュール
 
-- `libs.data.aggregate`: データ集計
 - `libs.data.modify`: DBデータ操作
 - `libs.data.search`: DB検索
 """
