+++ libs/domain/__init__.py
@@ -7,4 +7,5 @@
 - `libs.domain.stats`: 成績解析
 - `libs.domain.rule`: ルールセット管理
 - `libs.domain.placeholder`: プレースホルダ構築
+- `libs.domain.aggregate`: データ集計
 """
