+++ libs/domain/aggregate.py
@@ -1,5 +1,5 @@
 """
-libs/data/aggregate.py
+libs/domain/aggregate.py
 """
 
 from typing import Any, Optional
